+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Unreleased
 
+- [...]
 - **[UPDATE]** Add `background` prop to `CheckShieldIcon`
 - **[NEW]** New `ScamsEducationIllustration`
-- [...]
+- **[FIX]** Prevent `NewItinerary` from visually breaking with long names
 
 # v58.0.0 (25/03/2021)